websocket_client.py
#!/usr/bin/env python3
"""
Polymarket CLOB WebSocket Client
Connects to Polymarket CLOB WebSocket for real-time order book data
"""

import json
import time
import threading
import websocket
import requests
from datetime import datetime
from typing import Optional, Dict, List, Callable
from dotenv import load_dotenv
import os

load_dotenv()

class PolymarketCLOBWebSocket:
    def __init__(self):
        """Initialize Polymarket CLOB WebSocket connection"""
        self.WS_URL = os.getenv("CLOB_WS_URL")
        self.CLOB_API_BASE = os.getenv("CLOB_API_BASE")
        
        # WebSocket connection
        self.ws = None
        self.ws_thread = None
        self.connected = False
        self.running = False
        
        # Subscriptions
        self.subscribed_assets: List[str] = []
        self.subscribed_markets: List[str] = []
        
        # Order book data storage
        self.order_books: Dict[str, Dict] = {}  # asset_id -> {bids: [], asks: [], timestamp}
        
        # Callbacks
        self.on_book_update: Optional[Callable] = None
        self.on_price_change: Optional[Callable] = None
        self.on_trade: Optional[Callable] = None
        self.on_connect: Optional[Callable] = None
        self.on_disconnect: Optional[Callable] = None
        self.on_error: Optional[Callable] = None
    
    def on_message(self, ws, message):
        """Handle incoming WebSocket messages"""
        # TODO: Implement your own message parsing logic
        # Messages may arrive as JSON objects or arrays of events
        # Common event types: book, price_change, last_trade_price, ping/pong
        raise NotImplementedError(
            "WebSocket message handling removed. "
            "Implement your own message parser."
        )
    
    def _process_single_message(self, data):
        """Process a single message object"""
        # TODO: Route messages to appropriate handlers based on event_type
        # Event types include: book, price_change, last_trade_price, subscribed, error, ping
        raise NotImplementedError(
            "Message routing logic removed. "
            "Implement your own event type dispatcher."
        )
    
    def _handle_book_update(self, data: Dict):
        """Handle order book update"""
        # TODO: Parse order book data (bids/asks arrays)
        # Extract best bid and best ask prices
        # Store in self.order_books and call self.on_book_update callback
        raise NotImplementedError(
            "Order book parsing logic removed. "
            "Implement your own bid/ask extraction from book snapshots."
        )
    
    def _handle_price_change(self, data: Dict):
        """Handle price change event"""
        # TODO: Parse price_changes array from the message
        # Update self.order_books with new best bid/ask
        # Call self.on_price_change callback when both sides are available
        raise NotImplementedError(
            "Price change handling logic removed. "
            "Implement your own price update parser."
        )
    
    def _handle_trade(self, data: Dict):
        """Handle last trade price event"""
        if self.on_trade:
            self.on_trade(data)
    
    def on_error_handler(self, ws, error):
        """Handle WebSocket errors"""
        if self.on_error:
            self.on_error(error)
    
    def on_close_handler(self, ws, close_status_code, close_msg):
        """Handle WebSocket close"""
        self.connected = False
        if self.on_disconnect:
            self.on_disconnect(close_status_code, close_msg)
        
        # Auto-reconnect if still running
        if self.running:
            time.sleep(5)
            if self.running:
                self.connect()
    
    def on_open_handler(self, ws):
        """Handle WebSocket open"""
        self.connected = True
        
        self._start_keepalive()
        
        if self.on_connect:
            self.on_connect()
        
        # Resubscribe to previous subscriptions
        if self.subscribed_assets:
            time.sleep(1)
            self._subscribe(self.subscribed_assets, self.subscribed_markets)
    
    def _start_keepalive(self):
        """Start keepalive ping to maintain connection"""
        # TODO: Implement keepalive mechanism to prevent connection timeout
        raise NotImplementedError(
            "Keepalive logic removed. "
            "Implement your own ping/pong keepalive loop."
        )
    
    def _subscribe(self, asset_ids: List[str], market_ids: List[str] = None):
        """Send subscription message"""
        # TODO: Send subscription message to WebSocket
        # Format and send the proper subscription payload for the CLOB WS API
        raise NotImplementedError(
            "Subscription logic removed. "
            "Implement your own subscription message format."
        )
    
    def subscribe(self, asset_ids: List[str], market_ids: List[str] = None):
        """
        Subscribe to order book updates for specific assets
        
        Args:
            asset_ids: List of token/asset IDs to subscribe to
            market_ids: Optional list of market/condition IDs
        """
        if not isinstance(asset_ids, list):
            asset_ids = [asset_ids]
        
        if market_ids and not isinstance(market_ids, list):
            market_ids = [market_ids]
        
        return self._subscribe(asset_ids, market_ids)
    
    def get_order_book(self, asset_id: str) -> Optional[Dict]:
        """Get current order book for an asset"""
        return self.order_books.get(asset_id)
    
    def get_best_bid_ask(self, asset_id: str) -> Dict:
        """Get best bid and ask prices for an asset"""
        # TODO: Extract best bid/ask from stored order book
        raise NotImplementedError(
            "Best bid/ask extraction removed. "
            "Implement your own order book price extraction."
        )
    
    def get_order_book_summary(self) -> Dict:
        """Get summary of all order books"""
        # TODO: Build summary dict across all tracked order books
        raise NotImplementedError(
            "Order book summary logic removed. "
            "Implement your own summary builder."
        )
    
    def get_order_book_from_rest(self, token_id: str) -> Optional[Dict]:
        """Get order book from CLOB REST API"""
        # TODO: Fetch order book snapshot from REST endpoint
        raise NotImplementedError(
            "REST order book fetch removed. "
            "Implement your own REST API call to get order book data."
        )
    
    def get_price_from_rest(self, token_id: str) -> Optional[Dict]:
        """Get best bid and ask prices from CLOB REST API"""
        # TODO: Use get_order_book_from_rest and extract prices
        raise NotImplementedError(
            "REST price fetch removed. "
            "Implement your own REST-based price lookup."
        )
    
    def connect(self, debug: bool = False):
        """Connect to Polymarket CLOB WebSocket"""
        # TODO: Create WebSocketApp, start in a background thread, and wait for connection
        raise NotImplementedError(
            "WebSocket connection logic removed. "
            "Implement your own WebSocket connection setup."
        )
    
    def disconnect(self):
        """Disconnect from WebSocket"""
        self.running = False
        self.connected = False
        
        if self.ws:
            self.ws.close()
    
    def is_connected(self) -> bool:
        """Check if WebSocket is connected"""
        return self.connected
