requirements.txt
websocket-client==1.6.4
requests==2.31.0
py-clob-client>=0.34.5
python-dotenv
web3
python-dateutil

